bronze/pyproject.toml
[project]
name = "bronze"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "delta-spark>=4.0.0",
    "pyspark>=4.0.1",
]
